requirements.txt
alembic==1.0.10
aniso8601==6.0.0
asn1crypto==0.24.0
attrs==19.1.0
cffi==1.12.3
Click==7.0
cryptography==2.6.1
Flask==1.0.2
Flask-Cors==3.0.7
Flask-HTTPAuth==3.2.4
Flask-Migrate==2.4.0
flask-restplus==0.12.1
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
Mako==1.0.9
MarkupSafe==1.1.1
passlib==1.7.1
psycopg2-binary==2.8.2
pycparser==2.19
PyJWT==1.7.1
pyrsistent==0.15.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.3
Werkzeug==0.15.2
